requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Compress==1.14
Flask-DotEnv==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyodbc==5.2.0
python-dotenv==1.0.1
Werkzeug==3.1.3